Muevo funciones de cron fuera de helpers
