+++ ckanext/gobar_theme/helpers/helpers.py
@@ -10,7 +10,6 @@ from urlparse import urljoin
 from urlparse import urlparse
 
 import moment
-from crontab import CronTab
 from dateutil import parser, tz
 from pydatajson.core import DataJson
 from pylons import config as config
@@ -254,25 +253,6 @@ def get_andino_base_page():
 
 def get_default_series_api_url():
     return config.get('seriestiempoarexplorer.default_series_api_uri', '')
-
-
-def search_for_cron_jobs_and_remove(comment_to_search_for):
-    # Buscamos y eliminamos los cron jobs que contengan el comment especificado por parámetro
-    if comment_to_search_for:
-        cron = CronTab(get_current_terminal_username())
-        jobs_with_specified_comment = cron.find_comment(comment_to_search_for)
-        cron.remove(jobs_with_specified_comment)
-        cron.write()
-
-
-def create_or_update_cron_job(command, hour, minute, comment=''):
-    if comment:
-        search_for_cron_jobs_and_remove(comment)
-    cron = CronTab(get_current_terminal_username())
-    job = cron.new(command=command, comment=comment)
-    job.hour.on(hour)
-    job.minute.on(minute)
-    cron.write()
 
 
 def get_current_terminal_username():
